Literature Review Final Draft
Commit includes the final draft of the literature review which will be submitted in part of the whole dissertation, this includes a brief introduction to the general technology which will be explored and researched in the whole length of the paper. Subtopics are also listed which mention and describe the different technics and technologies HTTP Streaming is made up of. Each technology is elaborated upon from different research papers referenced from IEEE. References have been added at the bottom of the document. Diagrams with corresponding captions are added also added to the document. Recommendations given by my mentor were followed and include sourcing various papers which touch upon each different technology subsection, writing style to use and type of diagrams to use to successfully describe and explain the process at hand.