Dissertation Methodology Commit
This commit includes a dissertation methodology which is sectioned in two main subtopics. The first to include a comprehensive explanation of each section of the implementation. The implementation is made up of the different technologies explained in the Literature Review. The second section is compiled of the testing and result breakdown which are going to be run on the aforementioned implementation. All the documentation is compiled in the third person and in the past tense. Figures and captions were included for the implementation section while tables were included for the testing section. A separate section was also included at the end of the document containing the known limitations prior the start of testing and the compiling of data.