Abstracts, Introduction & Conclusion Commit
Commit includes the remaining chapters for the dissertation paper. The Abstract has not been written however this will be straight to the point summary of the research with important research specifications. A passive writing style was adhered to along with the present continues writing style was used for the further work and recommendations sections. The introduction commit includes the first draft and includes the definition of the problem, a clear description of personal motivation for the research. All research objectives were defined along with the chapters in the research paper. The Conclusion committed has been finalised and contains a summary of all the previous chapters and evaluation of the researcher’s findings together with future work and recommendations.